+++ components/admin/Previews/PackagePreview/PackagePreview.test.js
@@ -111,7 +111,7 @@ describe( '<PackagePreview />, if data === undefined is returned,', () => {
   it( 'renders ApolloError', async () => {
     const wrapper = mount( Component );
 
-    await wait( 10 );
+    await wait( 15 );
     wrapper.update();
 
     const apolloError = wrapper.find( 'ApolloError' );